[REFACTOR] Use ViewModel Hook and Adapter Design Pattern
